+++ test/cli/test-block.js
@@ -4,7 +4,7 @@
 const expect = require('chai').expect
 const repoPath = require('./index').repoPath
 const describeOnlineAndOffline = require('../utils/on-and-off')
-const ipfs = require('../utils/ipfs')(repoPath)
+const ipfs = require('../utils/ipfs-exec')(repoPath)
 
 describe('block', () => {
   describeOnlineAndOffline(repoPath, () => {
