+++ test/cli/test-bootstrap.js
@@ -4,7 +4,7 @@
 
 const expect = require('chai').expect
 const repoPath = require('./index').repoPath
-const ipfs = require('../utils/ipfs')(repoPath)
+const ipfs = require('../utils/ipfs-exec')(repoPath)
 const describeOnlineAndOffline = require('../utils/on-and-off')
 
 describe('bootstrap', () => {
