+++ test/cli/test-files.js
@@ -6,7 +6,7 @@ const repoPath = require('./index').repoPath
 const fs = require('fs')
 const path = require('path')
 const describeOnlineAndOffline = require('../utils/on-and-off')
-const ipfs = require('../utils/ipfs')(repoPath)
+const ipfs = require('../utils/ipfs-exec')(repoPath)
 
 describe('files', () => {
   describeOnlineAndOffline(repoPath, () => {
