+++ test/cli/test-version.js
@@ -5,7 +5,7 @@ const expect = require('chai').expect
 const pkgversion = require('../../package.json').version
 const repoPath = require('./index').repoPath
 const describeOnlineAndOffline = require('../utils/on-and-off')
-const ipfs = require('../utils/ipfs')(repoPath)
+const ipfs = require('../utils/ipfs-exec')(repoPath)
 
 describe('version', () => {
   describeOnlineAndOffline(repoPath, () => {
